[FIX] l10n_be_hr_payroll: fix time off allocation
In the demo data, the number of paid time off was displayed as 21, when it should be 20.
The issue was tracked to the statement used for assigning the working schedule to each employee.
Also the maximum time off possible is now 24 days, as stated by the law, for people working 6 days a week and 20 for people working 5 days a week.

task-5172492

closes odoo/enterprise#98228

Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>