[FIX] stock_barcode: free the mobile barcode button
On small screens (e.g., Zebra scanner, iPhone 4), the image overlaps the barcode scan button, making it impossible to click and scan new items.

Steps to reproduce:
-------------------
On a small screen size (Zebra scanner, Iphone 4,..)
* Open barcode app
* Go on  Operations -> Receipts
* Apply a filter so that the list becomes empty.
-> the image is over the barcode button and the text.

Observation:
------------
Our image (in div .o_view_nocontent) has position absolute, anchoring it to the top part of the screen where the search bar is located (in div .o_control_panel).
The issue is that in barcode there is another element where the button to scan is located (o_kanban_tip_filter), and since the image is absolutely positioned relative to the top rather than this element, it ends up overlapping the button and the text, preventing interaction.

Why the fix:
------------
The solution we chose is to change the position of the image to relative, for the image to not hide the text.
Another possible solution was to make the image (more specificaly .o_view_nocontent .o_nocontent_help) ignore all input, allowing them to go through the image and reach the button belown.

opw-5123880

closes odoo/enterprise#98522

X-original-commit: 0c2b6ded1475f2f57c1b22882004ab62545dec39
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>
Signed-off-by: Joseph Vankelegom (josv) <josv@odoo.com>