[IMP] account_iso20022: partner_id required on payment
When doing a payment with SEPA as the payment method, and then create a batch
payment out of it. It could happen that the partner_id of the payment was not
set. That would cause a traceback because in the _get_CdtTrfTxInf we do a browse
on the partner to use it later on. But since the partner is False, we have an
empty record set.

task-5213880

closes odoo/enterprise#101978

X-original-commit: c5671eee0379a1b74d07636e521eee55429935a6
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>