[FIX] delivery_dhl_rest: set default values for DHL pallets
## Short functional explanation of the error
When setting the 'DHLxw Package Type' of a DHL delivery method to
'DHL pallets' without changing default values, trying to validate
an order delivery causes an error.

## Reproduction Steps
1. Type 'Delivery method' in the dashboard. Select DHL BE,
set the region to Europe and the 'DHLxw Package Type' to 'DHL
pallets'. Click on the arrow at the right of 'DHL pallets'. You
should see the default values set to 1200, 1600 and 1200. These
correspond to millimeters, while DHL expects centimeters.
2. Make sure you're using a company located in Belgium.
3. In the inventory settings, make sure that either 'multi-step
route' is unchecked or that for your warehouse, the outgoing
shipments are set to '1 step'.
3. Create a new quotation. Select a Customer with a Belgian address.
Add a product for which there's at least 1 copy in stock.
4. Click on add shipping and select DHL BE. Then click add and ok.
5. Click on confirm. A smart button at the top of the page called
"delivery" should appear. Click on it and set the scheduled date to
a future day. Finally, click validate.

### Expected behavior
New labels should generate without issue.

### Unexpected behavior
An error occurs:
```Bad request Multiple problems found, see Additional Details
1001: The requested product(s) (N) not available based on your
search criteria.(687662d08999745d3df3783068fcca24)
410107: The piece height exceeds max, N/-
410304: No products available
```

## Origin of the issue
The default values for DHL are set to 1200, 1600 and 1200, which
correspond to millimeters. However, DHL expects centimeters, for
which the default value is thus too high.
Note: because the issue stems from default values, after
implementing the fix, the module should be reinstalled.

__
opw-4804639

closes odoo/enterprise#98319

X-original-commit: 760cd842051d1aeb6f40bb36ad46c0d63f31d4a8
Signed-off-by: Steve Van Essche <svs@odoo.com>