[FIX] mrp_mps: avoid extra POL in test_replenish with demo data
The `test_replenish` test was failing with demo data because replenishment
created an extra Purchase Order line. The `_run_buy` search domain included
`date_planned_mps` with an equality check on a datetime stored in `UTC`.

With demo data loaded in a non-UTC timezone (e.g. Europe/Brussels),
the forecast date was converted to `2025-07-31 22:00:00 UTC`,
which did not match the existing PO at `2025-08-01 00:00:00 UTC`.
As a result, no PO was found and a duplicate was created.

Changes:

Set the test user timezone to UTC so that `date_planned_mps` comparisons
are stable when using demo data.
This ensures replenishment reuses the existing PO instead of creating a duplicate.

[runbot-230425](https://runbot.odoo.com/odoo/error/230425)

closes odoo/enterprise#98268

Signed-off-by: Steve Van Essche <svs@odoo.com>