[FIX] marketing_automation: prevent error on removing interval type in activity
Currently, an error occurs when removing the interval type or validity duration
type in a marketing automation activity. This happens when the system tries to
compute the "Activity Summary", which expects valid selection values but
receives False.

Steps to Reproduce:
1. Install Marketing Automation.
2. Create a new campaign and click on "Add new activity".
3. In the "Trigger" field, remove the duration (e.g., "Hours").

Error:
KeyError - False

Cause:
At [1], it tries to get the label from the selection dictionary using the value
of interval_type or validity_duration_type. When either field is False, it
causes a KeyError.

Fix:
This commit prevents the error by;
- Setting an empty summary when required selection values are missing.
- Returning 0 for interval standardized when interval type is not set.

[1]: https://github.com/odoo/enterprise/blob/bb92db89d3fc1f654dfdff9b2e334b13d81cb7ab/marketing_automation/models/marketing_activity.py#L228

sentry-6957343166

closes odoo/enterprise#98367

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>