[FIX] l10n_cl_edi: limit search to one record for recipient company
Before this commit, if multiple companies shared the same RUT (VAT),
for example companies with branches, searching a recipient could
return more than one company record. This could lead to errors when
accessing variables that expect a single record.

To prevent this issue, the search is now limited to one company record.

OPW-5128543

closes odoo/enterprise#98420

X-original-commit: 199c814b838d4ce82e46f6eafe070e8a25f95974
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Victor Miguel Armenta Carrillo (vmac) <vmac@odoo.com>