[FIX] planning: process task assignation with multicompany
_______________________________________
Let's say we have the scenario where a user has access to
company_1 and company_2, but only has a corresponding
employee in company_2. If the user selects company_1 and
company_2 but keeps company_1 as his current company, and
tries to assign himself a task that has been created for
company_2, nothing happens.

1. As an admin, create a user with which you'll be able to
log. Make sure that you have at least 2 companies created,
and that the user has access to both. Create an employee
for that user in company_2.
2. Select both companies. In planning, create a slot for
company_2 and publish it.
3. Log in as the user you created. Make sure that the
current company is company_1. Select company_2.
4. Go to planning and try to assign yourself to the slot
you've just created as an admin

Either an error message shows, or the employee is assigned
to the slot for company_2 (as company_2 is selected).

Nothing happens

When the current company isn't the one corresponding
to the one the employee is in, even if another company
is selected and contains the employee,
self.env.user.employee_id is set at False

_________________________________________
opw-4963674

---

closes odoo/enterprise#95933

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Maeva de Keyser (mdek) <mdek@odoo.com>