[FIX] website_helpdesk_knowledge: redirect to knowledge article
To reproduce:
=============
1. Create a Helpdesk Team linked to a Knowledge Article
2. Access the Help page on website as a portal user
3. Click on "Browse Articles" button
-> redirected to empty knowledge home portal view

Problem:
========
before this commit, redirection was made through the method
`redirect_to_article` which will later call `_redirect_to_portal_view`
that doesn't use the `article` parameter anymore as there is a patch
on the front side to handle the redirection to the articale through
the router, but as the calls are server-side, the patch is not
applied and the redirection fails.

Solution:
=========
instead of calling `redirect_to_article`, directly redirect to
the article's `website_url`.

opw-5114885

closes odoo/enterprise#98597

X-original-commit: cade07c9213feeb637b1b703a2c55379dbef535a
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Abdelouahab Laaroussi (abla) <abla@odoo.com>