[FIX] test_discuss_full_enterprise: fix failing ai patches test
Purpose:
--------
Commit [1] added a test tour relying on the ai_app app which is not a
dependency of test_discuss_full_enterprise (only ai is).
This commit replaces usages of the views of the ai_app by the views of
im_livechat and crm which are dependencies of the module.

Fixes https://runbot.odoo.com/odoo/runbot.build.error/233603

[1]: https://github.com/odoo/enterprise/commit/a7b6eb985e85477f38f15c0e269c2b5ae10d37f6

closes odoo/enterprise#98371

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>