[FIX] ai: prevent error when user timezone is not set
Currently, an error occurs when opening the AI chat from the mail composer if
the user's timezone has been removed.

Steps to Reproduce:
1. Install AI and Purchase modules.
2. In the user's profile, remove the 'TimeZone'.
3. Open any purchase order and click on "Send PO".
4. In the "Compose Email" form, click on AI icon.

Error:
AttributeError - 'bool' object has no attribute 'upper'

Cause:
The issue occurs because self.env.user.tz returns False when the user's timezone
is not set, causing an error when it tries to use upper() method on bool value.

Fix:
This commit fixes the issue by defaulting to the UTC timezone when the
user's timezone is not set.

Ref: https://github.com/odoo/odoo/blob/0f40ea82a1332f0e7168d861ad446b7316ab03de/odoo/addons/base/models/res_partner.py#L225-L228

sentry-6961379038

closes odoo/enterprise#99039

X-original-commit: adf9da9d88d5ced19b7ca75bf69b49b9bb132729
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Aditi Patel (adip) <adip@odoo.com>