[ADD] l10n_us_direct_deposit: support direct vendor payouts
Direct Deposit is a common requirement in the US. Payment Initiation via
Ponto is not a good solution as that is only usable for Europe, so other
solutions have to be developed. Unlike Ponto via account_online_payment,
Wise is implemented directly in enterprise instead of odoofin as
customers can open an account themselves and handle all of the data
without Odoo being responsible for anything.

This module works by overriding the validation steps for batch payments
to utilize Wise's API to create a batch group on their side, before
redirecting to their final step where the customer can pay however they
choose. For the United States this can be from their account balance,
ACH transfer, fedwire transfer, or via direct deposit if they set that
up in their portal.

task-5080414

closes odoo/enterprise#96588

Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>