[FIX] l10n_mx_hr_payroll: Add demo company.
In the pr #95221 the data from the demo company where no new fields were added was deleted. This decision was made to standardize the demo data in the smallest number of modules. However, the l10n_mx_hr_payroll module does not depend on any other l10n_mx module, as it does not need any of the information added by the other modules, but it does add information about departments and other items based on the company's demo data. This caused an error when installing this module alone, as it attempted to add this information to a company that did not exist. To fix this, the company information is also added as demo in this module, as it was before.

I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr

closes odoo/enterprise#98404

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>