[IMP] l10n_br_avatax: send delivery address for nfse
Purpose: In Brazil, it is common for companies to sell
services in one city and provide the services in
another city, thus, it is necessary to inform the place
of service provision in NFS-e. Users can specify where
the service was provided through the delivery addresss
on either sale order or invoice. The delivery address
will be sent in the request to the tax calculation and edi.
The returned EDI document will display the city the service
was provided.

Outline of additional attributes being sent:
-  header.locations.rendered.address.street --> partner_shipping_id.street
-  header.locations.rendered.address.neighborhood --> partner_shipping_id.street2
-  header.locations.rendered.address.zipcode --> partner_shipping_id.zip
-  header.locations.rendered.address.cityName --> partner_shipping_id.city
-  header.locations.rendered.address.state --> partner_shipping_id.state_id.code
-  header.locations.rendered.address.countryCode --> partner_shipping_id.country_id.l10n_br_edi_code

task-5124608

closes odoo/enterprise#99214

Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>