[FIX] spreadsheet_edition: Allow users to delete compromised datasources
Currently, within a spreadsheet with broken datasources (i.e. with an
invalid model), the user can open the datasource sidepanel and only has
a message explaining the datasource is invalid. On the pivot sidepanel,
there is also a "delete" button to encourage the user to clean up their
spreadsheet. The encouragement is quite necessary as the broken
datasource will prevent a proper usage of the spreadsheet and notably
the ability to edit/create new global filters.

This revision adds the 'delete' button to all types of datasources
(i.e. charts & lists).

task-5220489

Part-of: odoo/enterprise#100989
Related: odoo/odoo#238155
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>