[FIX] account_followup: send attachments from template
Issue:
We can add attachments and dynamic report to the email templates, but they are not sent with follow-ups.

Steps to reproduce:
- Go to the "Payment reminder" email template
- Under the page "Content", add an attachment by clicking the "Attachments" button
- Under the page "Settings", add a dynamic report
- Create an overdue invoice for a partner
- Go on the form view of the partner, "Accounting" page
- Click "Send", make sure the template used is the one with the attachments
- Send
- The attachments on the template and the dynamic report are ont sent

Cause:
The mail template to send the follow-ups is only used to prefill the wizard.

Solution:
Add the template in `_get_wizard_options()` to add the template in the option and later use it to add/generate its attachments.

This commit also refactors how the attachments are computed:
The previous code was adding the invoices PDFs then removing them. The whole process was confusing.
Now `_get_followup_attachments()` returns the list with the necessary attachments depending
on the different options.

opw-5147736

closes odoo/enterprise#99586

X-original-commit: 4d15762580868c93d48ad316389ba01a187302ac
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>