[FIX] l10n_be_reports,_client_nihil: Bring back client nihil option
Since the new return system, the client nihil option that was inside the vat export xml was removed. We used a too simple mechanism which only checks if the tax report was empty or not.

This was completely wrong, now we added the checkbox back for finer tuning and we precomputed it.
The precompute is checking if any partner is exceeding 250€ invoiced for the current year.

This commit adds a new module l10n_be_report_client_nihil which will be removed in master. It will be integrated with l10n_be_reports directly.

task-5217079

closes odoo/enterprise#99496

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>