[PERF] l10n_fr_fec_import: Avoid OOM when importing FEC
Description:

When importing an FEC, Odoo will fetch all the partners to link the new
imported records to the existing partners. The issue is that it triggers
the prefetching of all the fields of the partners (304k partners in their
case), causing a memory error. To avoid that, we can just fetch the field
that we need (e.g. "name" and "ref").

Reference:

opw-5153555

closes odoo/enterprise#99618

X-original-commit: 188a6ec0670a686d16dc6e2e8b26898109768281
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>