[FIX] stock_barcode{,_mrp}: replace lot on barcode scans if unscanned
Issue:

Scanning an unreserved lot of a kit component creates a new barcode line
rather than updating the value of the initially reserved lot. As a
result, the backorder porcess considers that every unscanned yet
initially reserved quantity is to backorder.

Steps to reproduce:
- Create a kit product with a kit BOM:
    - 1 x COMP (tracked by SN)
- Add two Serial numbers SN001 and SN002 in stock for the COMP product
- Create and confirm a delivery order for 1 unit of oyur kit product
- Go the barcode app to process your delivery
- Scan SN002
> A new line is created instead of updating the initial reservation
- Validate the delivery
> A backorder dialog opens proposing to update the unscanned reservation

Cause of the issue:

Scanning a lot will first try to find a line to update, however,
currently a line will only be found if the scanned lot has been reserved
or if no particular lot has been reserved:
https://github.com/odoo/enterprise/blob/a1113eebf634302f7ceece612e7ee068c99781e4/stock_barcode/static/src/models/barcode_model.js#L1659-L1661
https://github.com/odoo/enterprise/blob/a1113eebf634302f7ceece612e7ee068c99781e4/stock_barcode/static/src/models/barcode_model.js#L743-L746
In particular, since no line is considered as valid, a new line is
created. And, since this new line does not refer to any `move_id` while
the existing one does,  the move with the initial reservation will be
backordered considering none of its demand was fulfilled:
https://github.com/odoo/enterprise/blob/a1113eebf634302f7ceece612e7ee068c99781e4/stock_barcode/static/src/models/barcode_picking_model.js#L904-L921

Fix:

In order to loosen the condition of lot override on barcode lines we add
a check on the package and the location of the line in order to avoid
use cases where the initial move line already contains info's that are
proper to the initial lot.

opw-5100026

closes odoo/enterprise#99916

X-original-commit: 5e2820449359cf8a48c68ffad5d16bba3e5fd6b6
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>