[FIX] documents: fix documents child_of
`user_folder_id` and `folder_id`'s `child_of`
were not taking into account that the path could be broken if
inaccessible folders are between accessible documents.

E.g., the DB structure

COMPANY
└── Folder A
    └── Folder B (inaccessible to User A)
        └── Folder C

Would appear to user A as

COMPANY
└── Folder A
SHARED
└── Folder C

such that "Folder C" (and descendants) should not be found
when searching `child_of` "Folder A" or "In Company".

Note that more creativity would be necessary to fix this before 19.0
as the stored `folder_id` field could not be `_search`ed.

Performance evaluations resulted in choosing this version over a few
others (using regular user on our own data).

Task-5231269

closes odoo/enterprise#99040

Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>