[FIX] project_enterprise: prevent error when computing the dependency warning
Currently, an error occurs when computing the dependency warning for tasks.

Error:
    Expected singleton: project.task(8, 7).

After this commit if task1 depends on task2, if task1's task dependency is
False, then the dependency warning for task1 should also be False.

However, when task dependency is enabled and two or more tasks are blocked
by a single task, changing the Deadline of the blocking task triggers the
computation of the dependency warning for those dependent tasks. During this
process, the system attempts to check the task dependency across multiple
records, which raises the error at [1].

This commit ensures that, since self may contain multiple tasks, the dependency
warning is computed safely for each specific task.

[this commit]: https://github.com/odoo/enterprise/commit/50768627accb9d9767514fda7c6975bf4d8227d3#diff-b4bb763881453b0d69cae371c4c5034f6e199dfaf3a3056aba3381cf64dbf3dc

[1]: https://github.com/odoo/enterprise/blob/769201e22618b6683863f0839d7ad80bea0927cf/project_enterprise/models/project_task.py#L350

sentry-6914184933

closes odoo/enterprise#98532

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>