[FIX] pos_urban_piper: fix error when making test order
When creating a test order for a product whose tax is configured
as a Group of Taxes without any definitions, a traceback occurs.

Steps to reproduce the error:
- Install "pos_urban_piper" module
- Configure Urban Piper integration for the PoS
- Create a new tax without Definition > Tax Computation: Group of Taxes > Save
- Create a new product > Set the above tax in sales taxes > Save
- Open the session for PoS
- Go to Settings > Urban Piper Location > Set Food Delivery Platforms > Save
- Click Test Order > Select the product and Delivery Provider > Place Order

Traceback:
"IndexError: list index out of range"

https://github.com/odoo/enterprise/blob/92bb923ffe185b7744adeadcc8f2972f9a64effb/pos_urban_piper/controllers/main.py#L319-L322
The issue happens because "flatten_taxes_hierarchy()" calls
"_flatten_taxes_and_sort_them()" method at [1],
which returns an empty record when the group tax has no children at [2].
As a result, "tax_types" becomes an empty list ([]), leading to an IndexError
when trying to access "tax_types[0]".

[1]: https://github.com/odoo/odoo/blob/57850b32332a71933a4f6b52d7428684de831f4b/addons/account/models/account_tax.py#L2786

[2]: https://github.com/odoo/odoo/blob/57850b32332a71933a4f6b52d7428684de831f4b/addons/account/models/account_tax.py#L795

sentry-6984648461

closes odoo/enterprise#100958

X-original-commit: 2f2b53f1c6f31ae22351d22cf4bf59ef01a63691
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>
Signed-off-by: Renilkumar Kajavadra (reka) <reka@odoo.com>