[FIX] account_budget: access error with unrestricted analytic account
Steps to reproduce:
- Create an analytic account [ACCOUNT] not restricted to any company
- Create an analytical distribution model applying [ACCOUNT] to [PRODUCT]
  not restricted to any company
- Create a budget in company [COMP A] using analytic account [ACCOUNT]
- Create a budget in company [COMP B] using analytic account [ACCOUNT]
- Select only company [COMP B]
- Create a purchase order, add [PRODUCT] to a line

Issue:

Access Error is raised
```
[...]
Sorry, Mitchell Admin (id=2) doesn't have 'read' access to:
- Budget Line, Budget Your Company (budget.line: 8, company=YourCompany)

Blame the following rules:
- Budget lines multi-company
[...]
```

Analysis:

Because we search with sudo company rules are bypassed and the system
assign both budget lines to the order line (budget_line_ids).
When we access the field, the system try to access the budget line
of the other company, causing the error.

As solution we revert 4941c1928c9a2bcb38af3cadf023f904c9fd9be0
removing the `sudo` on the search and we explicitly check if the user
has read access on the budget analytic model

opw-5056712

closes odoo/enterprise#99155

X-original-commit: da2029e828f1db1e368b37c55360d89c53ba5c76
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>