[FIX] test_discuss_full_enterprise: fix ai test 'hr.employee' missing
HOOT test "can handle command and disable mentions in AI composer"
fails non-deterministically on runbot with the following error:

```
Cannot find a definition for model "hr.employee": could not get model from server environment (did you forget to use `defineModels()?`)
```

Somehow this test requires `hr.employee` model, which explained why
it had `defineHrModels()`. This was replaced to enterprise livechat
models, since new test need livechat and `hrModels` lacked livechat.

The way to `defineModels()` on modules with lots of dependency sucks:
`hr` and `im_livechat` and `ai` have no explicit dependency among
each other, but a module such as `test_discuss_full_enterprise` has
to combine all of their behaviors and definitions. `defineModels()`
is designed to extend a model by makin a subclass of a definition
of same model in a dependent module, which is not practical here.

This commit doesn't solve the problem of defining livechat and hr
models in the `test_discuss_full_enterprise` test suites, but for the
quick and immediate fix it moves the problematic AI test in its own
suite with `defineHrModels()`.

Fixes https://runbot.odoo.com/runbot/build/92414152

closes odoo/enterprise#98642

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>