[FIX] l10n_de_reports: fix datev export of bill with edited tax amount
Steps to reproduce:
- Install Accounting and l10n_de_reports
- Switch to a German company

- Create a bill:
  * Price: 100.00
  * Taxes: 19%
- Edit the tax total with the pencil button

- Go to "Accounting / Reporting / Audit Reports / General Ledger"
=> The tax amount is the one that has been edited manually
- Download "Datev DATA (zip)"
- Open "EXTF_accounting_entries.csv" file

Issue:
The total amount in the file is the one before the edition of the
tax amount.

Cause:
The Datev data depends on "price_total" field of the invoice lines,
but this field is not updated when the tax amount is edited manually.

Solution:
Check the total by adding "price_total" of each invoice line and the
total amount defined in "tax_totals" field.
If there is a difference, compute the delta for each tax group and
split it between all the lines where a tax of that group is used.

opw-4951488

closes odoo/enterprise#100244

X-original-commit: b7822d77c63eb4bec7f3c9e3109e8180e47c5bd5
Signed-off-by: Paolo Gatti (pgi) <pgi@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>