[FIX] delivery_easypost: Load EasyPost carriers without traceback
**Issue**
A traceback occurs whenever a carrier is selected in the EasyPost delivery method.

**Steps to reproduce**
1. Install the barcode application
2. Go to Settings > Inventory > enable "Easypost" shipping connector
3. Go to Delivery Methods -> click "New" -> choose "Easypost" as provider
4. Fill in the API keys
5. Click "Load your Easypost carrier accounts" and select a carrier
→ A traceback occurs

**Cause**
The error occurs at:
https://github.com/odoo/enterprise/blob/39320569a3561f133000ded00800f705d9507ac2/delivery_easypost/static/src/components/carrier_type_selection/carrier_type_selection.js#L15

`ev` is expected to be a JSON object but is received as a string.
The component extends `SelectionField`, which has been modified in [this commit](https://github.com/odoo/odoo/commit/0923409082ead5ffdf2c28f3b063fbd91ebce553):
https://github.com/odoo/odoo/blob/5be6ef1ae3958a25e680f5fa30824fec57d63efe/addons/web/static/src/views/fields/selection/selection_field.xml#L9

**Solution**
Adapt the `onChange` method to handle the argument type change.

opw-5170833

closes odoo/enterprise#98678

Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>