[FIX] sale_subscription_stock: update delivered qty on return
**Steps to reproduce**
- Create a new subscription using a subscription product. Confirm
it.
- Validate the generated delivery.
- Return the delivery and validate the return.
- Issue: the delivered quantity of the sale order line is not updated.

**Cause**
Currently, we consider a move as related to a subscription period based
on the `date_deadline` field (see _get_outgoing_incoming_moves).
Since `_prepare_procurement_values` is not called when creating a return,
the `date_deadline` is not set on the return moves.

**Solution**
Set the `date_deadline` on subscription return moves based on the original
move.

opw-5136406

closes odoo/enterprise#100097

X-original-commit: 0dd551402018d782c256b128b0e2b3ca746331aa
Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>