[FIX] sale_stock_renting: fix rental virtual available with partially returned rental order
Issue:

Due to this bug, virtual available is not calculated correctly
when there are rental orders with some returned lines.

To reproduce:
1- Create a storable rental product with 10 units in stock.
2- Enable the 'Rental Transfer' setting.
3- Create a rental order with two lines:
 line 1: 4 units
 line 2: 1 unit
4- Confirm and pick up order lines
5- Return the line with 1 unit
6- Create another order: it must show 6 available but 7 is shown

Cause:

This is caused due to filter on lines in
`_get_virtual_unavailable_qty_in_rent` which filters lines
with `order_id.rental_status`. However this filter misses
the lines which have different status than the `order_id`.
In #88689, `rental_status` field is added to `sale_order_line`.
Using the `sale_order_line.rental_status`, we can ensure only
line with status in `return` and `pickup` are considered.

opw-5185084

closes odoo/enterprise#98689

Signed-off-by: Mohammadmahdi Alijani (malj) <malj@odoo.com>