[FIX] l10n_it_pos: fix printing reward lines
This PR fixes the case of printer blocked using reward lines
and multiple order lines.
To reproduce the issue:
- Add a product with reward (for example 10% reward on 50€ , that is 5€)
- Add a second product without reward that costs less than 5€
- Print the receipt on fiscal printer
The lines are orderer before the non reward lines, than reward lines.
The printer is blocking because the adjustmentType discounts last line
of less than 5€. It is wrong.
Another way to reproduce the issue:
- Add a product with reward
- Add a second product with reward
- The reward line with the sum of both rewards is created
The same error appear because it discounts both lines.
It is better to discount the entire department instead of the last sale

closes odoo/enterprise#101171

X-original-commit: 1bc66c82d8cdf24eeb7f3c2dc892593106464b31
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>
Signed-off-by: Paolo Gatti (pgi) <pgi@odoo.com>