[FIX] esg: fix related fields not updating
The `currency` and `unit` fields on emitted emissions are supposed to be related fields on the emission factor.
However, due to how the report combining accounting emissions and other emissions is implemented,
they are not correctly updated when the factor is modified.

This commit makes them computed fields and enable the "store" attribute en them.
Despite there being no actual table to store data into, this allows us to do a round trip to the server to fetch the correct values.

closes odoo/enterprise#98737

X-original-commit: 42ac10ad9177dc0a73dc308b79ba378bd1e6e5a0
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>