[IMP] voip: use country data from res.country model
Previously, we hard-coded the URL according to the country code to
display a country flag. However, this is not the case for all countries.
Countries like Bonaire, Sint Eustatius, and Saba use the flag of the
Netherlands. Additionally, Bonaire, Sint Eustatius, and Saba are shown
as Caribbean Netherlands, which is inconsistent with the data from our
res.country model.

This commit changes it to use the data from res.country.

Task-5207454

closes odoo/enterprise#99136

X-original-commit: f48efdd16c48377b68ea7063c19b2c9da3dcd934
Signed-off-by: Yuchen Huang (yhu) <yhu@odoo.com>