[FIX] stock_barcode(_mrp): ensure correct key is set in parsed result
This error occurs while entering the barcode number.

Steps to reproduce:
- Install 'stock_barcode' module(without demo)
- Setting > Barcode > Set Barcode Nomenclature = 'Default GS1 Nomenclature'
- Barcode > click on 'Scan or tap' > Enter a barcode(barcode number should
  startwith 'urn' (eg: "urn:epc:tag:sgtin-96 : 3.0614141.038656.0")) > Apply

Traceback:
KeyError: 'rule'
AttributeError: 'list' object has no attribute 'get'

At [1], we expect the key 'rule' to be present in the result, but this key is
never set in the return statement at [2]. Since [1] also relies on the result’s
type, a new key value 'type' has been added in this [commit] to address that.

This commit ensures that the correct keys are passed in the result dictionary.

[1]: https://github.com/odoo/enterprise/blob/b001e9cc2af0f800e2a7965b61aa9b9c5bd4e89e/stock_barcode/controllers/stock_barcode.py#L29-L31

[2]: https://github.com/odoo/odoo/blob/f8f72b15598576f5870e49879e96fc5c127a6100/addons/barcodes/models/barcode_nomenclature.py#L174-L189

[commit]: https://github.com/odoo/odoo/commit/1394fa161a6fcf77b4443cbf784ad7dd635e7f9e#diff-be2a58d0591614180295c070396ff487f4bc04bf33aad2829d7bfab4671a792cR60

sentry-6992944243

closes odoo/enterprise#99583

X-original-commit: 3dc7d21ac96c0971daf2eb9167044694fac6de8f
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Aayush Modi (aamo) <aamo@odoo.com>