[FIX] account_accountant: handle installments in reco
From a previous fix https://github.com/odoo/enterprise/commit/c61e12df64904c975b9abd1a6ba0ec359f06874d
The flow was changed from what it used to be.
When in the bank reconciliation widget, we want to add all the selected
move lines, even if sum of their amounts exceed the statement line's
one.
However, when adding the statement line as outstanding debit/credit
from the invoice/bill view form, we want the reconciliation to stop
whenever the whole statement line amount is consumed, and create a
a partial if needed.

Thus, the previous commit is mainly reverted in favour of a per-case
handling of the reconciliation.

Part-of: odoo/enterprise#100533
Related: odoo/odoo#237421
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Pierre-Rodéric Roose (prro) <prro@odoo.com>