[FIX] stock_barcode: default inventory location
Since previous commit, the barcode's tests use a company dedicated to
the testing environment. While this solution has undoubted benefits, it
comes with one flaw for the tests: if `mrp_subcontracting` is installed,
the subcontracting location will be created before the stock location.
In the Barcode app, defualt location for inventory adjustments are
either the last scanned location, either the first fetched location.
In this case, that means it will be the subcontracting company, since
locations are stored in the barcode lazy cache by their ID and the
subcontracting location has smaller ID than the storage location ID.

To fix that, we put in JS cache only the warehouse stock location and
its sub-locations, and quants' locations if it's relevant.

Part-of: odoo/enterprise#96657
Related: odoo/odoo#234483
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>