[IMP] l10n_gt_edi: send xml along with pdf in customer email for e-invoices
Purpose: In the Guatemalan localization, when an electronic invoice is created,
the email template to the customer already contains the PDF version of the DTE.
However, the XML file, generated and sent to the SAT through Infile, is not
attached to that email. Even though the SAT and Infile deliver the XML to the
customer, it is more convenient and transparent if Odoo includes it directly in
the outgoing invoice template email, so the customer receives both files in one
place.

Before this commit:-
- Only PDF version is attached by default in customer email
for e-invoices.
- Name of edi document is prefixed with `Demo` if company is in demo environment
(check parent company's environment in case of child company)
(e.g. DEMO_certificate_INV_2025_00001.xml) and prefixed with `SAT` if company is
in testing or production environment(e.g. SAT_certificate_INV_2025_00001.xml).

After this commit:-
- XML version is also attached by default along with PDF in customer
email for e-invoices.
- Name of edi document is always prefixed with `SAT`.

task-5224521

closes odoo/enterprise#102322

X-original-commit: e06b5c68449fffed58ae1e9d9dc9b7acc8e2031c
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Pratham Nareshbhai Patel (ppat) <ppat@odoo.com>